+++ app/Http/Controllers/Api/V1/LoanContributionAdjustController.php
@@ -144,61 +144,4 @@ class LoanContributionAdjustController extends Controller
         }
         return $adjust_contribution;
     }
-    /**
-    * Actualizar o Crear el array de Objeto Registro de Garantía de préstamos
-    * @bodyParam affiliate_id integer ID required de afiliado. Example: 9              
-    * @bodyParam guarantees[0].id  integer required ID del registro de la tabla loans o prestamos.Example: 1
-    * @bodyParam guarantees[0].type string required registro del nombre de las tablas tabla loans,prestamos. Example: prestamos
-    * @bodyParam guarantees[0].type_affiliate enum required (lender,guarantor,cosigner) Example:guarantor
-    * @bodyParam guarantees[0].quota numeric required cuota del prestamo de los garantes Example: 10000.50
-    * @bodyParam guarantees[0].type_adjust enum required (loan_guarantee_register)  Example: loan_guarantee_register
-    * @bodyParam guarantees[1].id  integer required ID del registro de la tabla loans,prestamos.Example: 1
-    * @bodyParam guarantees[1].type string required registro del nombre de las tablas tabla loans,prestamos. Example: prestamos
-    * @bodyParam guarantees[1].type_affiliate enum required (lender,guarantor,cosigner) Example:guarantor
-    * @bodyParam guarantees[1].quota numeric required cuota del prestamo de los garantes Example: 10000.50
-    * @bodyParam guarantees[1].type_adjust enum required (loan_guarantee_register)  Example: loan_guarantee_register
-    
-    * @authenticated
-    * @responseFile responses/loan_contribution_adjust/updateOrCreateLoanGuaranteeRegister.200.json
-    */
-    public function updateOrCreateLoanGuaranteeRegister(Request $request) { 
-      // return $request;
-        $request->validate([
-          'affiliate_id' => 'required|integer|exists:affiliates,id',
-          'guarantees' => 'required|array'
-        ]);
-        //return $request->guarantees;
-       $loan_contribution_guarantee_register_ids = collect();
-        foreach ($request->guarantees as $loan_contribution_adjust) {
-            $adjustable_type = $loan_contribution_adjust['type'];
-           // return $loan_contribution_adjust['id'];
-            $loan_contribution_guarantee_register = new LoanContributionAdjust();
-            $loan_contribution_guarantee_register->user_id = Auth::id();
-            $loan_contribution_guarantee_register->adjustable_id = $loan_contribution_adjust['id'];
-            $loan_contribution_guarantee_register->adjustable_type = $loan_contribution_adjust['type']="SISMU" ? "prestamos":"loans";
-            $loan_contribution_guarantee_register->affiliate_id = $request->affiliate_id;
-            $loan_contribution_guarantee_register->amount = $loan_contribution_adjust['quota'];
-            $loan_contribution_guarantee_register->description = $loan_contribution_adjust['code'];
-            $loan_contribution_guarantee_register->type_adjust = "loan_guarantee_register";
-            $loan_contribution_guarantee_register->type_affiliate = "guarantor";   
-            $loan_contribution_guarantee_register->period_date = Carbon::now()->format('d-m-Y');
-            $loan_contribution_guarantee_register->database_name = $adjustable_type;
-            $adjust_contribution = LoanContributionAdjust::where('affiliate_id',$request->affiliate_id)
-                                                  ->where('adjustable_type',$loan_contribution_guarantee_register->adjustable_type)
-                                                  ->where('adjustable_id',$loan_contribution_guarantee_register->adjustable_id)
-                                                  ->where('type_affiliate',$loan_contribution_guarantee_register->type_affiliate)
-                                                  ->where('type_adjust',$loan_contribution_guarantee_register->type_adjust)
-                                                  ->where('period_date',$loan_contribution_guarantee_register->period_date)
-                                                  ->whereNull('loan_id')->first();                             
-            if($adjust_contribution){
-                $adjust_contribution->update();
-                $loan_contribution_guarantee_register_ids->push($adjust_contribution->id);
-            }else{
-               $loan_contribution_guarantee_register->save();
-               $loan_contribution_guarantee_register_ids->push($loan_contribution_guarantee_register->id);
-            }                   
-        }
-        $loan_contribution_guarantee_register_ids= array('loan_contribution_guarantee_register_ids'=>$loan_contribution_guarantee_register_ids);
-        return $loan_contribution_guarantee_register_ids;
-    }
 }
